chore: improve webpack config
