+++ webpack.config.js
@@ -1,4 +1,4 @@
-const getConfig = require('./webpack.config.base');
+const { getConfig, dev } = require('./webpack.config.base');
 const { spawn } = require('child_process');
 
 let electronProcess;
@@ -6,33 +6,33 @@ let electronProcess;
 const mainConfig = getConfig({
   target: 'electron-main',
 
-  mode: 'development',
-
-  watch: true,
+  watch: dev,
 
   entry: {
     main: './src/main',
   },
 
-  plugins: [
-    {
-      apply: compiler => {
-        compiler.hooks.afterEmit.tap('AfterEmitPlugin', () => {
-          if (electronProcess) {
-            electronProcess.kill();
-          }
-
-          electronProcess = spawn('npm', ['start'], {
-            shell: true,
-            env: process.env,
-            stdio: 'inherit',
-          })
-            .on('close', code => process.exit(code))
-            .on('error', spawnError => console.error(spawnError));
-        });
-      },
-    },
-  ],
+  plugins: [],
 });
+
+if (dev) {
+  mainConfig.plugins.push({
+    apply: compiler => {
+      compiler.hooks.afterEmit.tap('AfterEmitPlugin', () => {
+        if (electronProcess) {
+          electronProcess.kill();
+        }
+
+        electronProcess = spawn('npm', ['start'], {
+          shell: true,
+          env: process.env,
+          stdio: 'inherit',
+        })
+          .on('close', code => process.exit(code))
+          .on('error', spawnError => console.error(spawnError));
+      });
+    },
+  });
+}
 
 module.exports = [mainConfig];
