+++ webpack.config.renderer.js
@@ -1,0 +1,52 @@
+const webpack = require('webpack');
+const { getConfig, dev } = require('./webpack.config.base');
+const { join } = require('path');
+const { execSync } = require('child_process');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
+
+const PORT = 4444;
+
+const config = {
+  target: 'electron-renderer',
+  plugins: [],
+  output: {},
+};
+
+if (dev) {
+  config.devServer = {
+    contentBase: join(__dirname, 'dist'),
+    port: PORT,
+    hot: true,
+    inline: true,
+    historyApiFallback: {
+      verbose: true,
+      disableDotRule: false,
+    },
+  };
+
+  config.output.publicPath = `http://localhost:${PORT}/`;
+
+  config.plugins.push(new webpack.HotModuleReplacementPlugin());
+}
+
+const appConfig = getConfig(config, {
+  entry: {
+    app: ['./src/renderer/views/app'],
+  },
+
+  plugins: [
+    new HtmlWebpackPlugin({
+      title: 'Wexond',
+      template: 'static/pages/app.html',
+      filename: 'app.html',
+    }),
+    new HardSourceWebpackPlugin(),
+  ],
+});
+
+if (dev) {
+  appConfig.entry.app.unshift('react-hot-loader/patch');
+}
+
+module.exports = [appConfig];
